+++ .storybook/main.ts
@@ -22,6 +22,12 @@ const config: StorybookConfig = {
     options.presets!.push('@emotion/babel-preset-css-prop');
     return options;
   },
+  staticDirs: [
+    {
+      from: '../src/assets/fonts',
+      to: 'src/assets/fonts'
+    }
+  ]
 };
 
 export default config;
