+++ src/styles/typo.stories.tsx
@@ -1,7 +1,5 @@
 import { type Meta } from '@storybook/react';
 
-import { pretendard } from '~/assets/fonts/pretendard';
-
 import { BODY_1, BODY_2_BOLD, BODY_2_REGULAR, HEAD_1, HEAD_2_BOLD, HEAD_2_REGULAR } from './typo';
 
 const meta: Meta<typeof Default> = {
@@ -13,7 +11,7 @@ export default meta;
 
 export function Default() {
   return (
-    <div className={pretendard.className}>
+    <>
       <p>HEAD 1</p>
       <p css={HEAD_1}>동료의 익명 피드백으로 완성시키는 나의 커리어 브랜딩 서비스</p>
 
@@ -31,6 +29,6 @@ export function Default() {
 
       <p>BODY 2 REGULAR</p>
       <p css={BODY_2_REGULAR}>동료의 익명 피드백으로 완성시키는 나의 커리어 브랜딩 서비스</p>
-    </div>
+    </>
   );
 }
